Align with tslint settings
fixup! Align with tslint settings
